// src/features/events/common/model/flowEvent.schema.ts
export const flowEventsSchema = z.object({
  Alert: z.array(alertEventSchema).optional(),
  Anomaly: z.array(anomalyEventSchema).optional(),
  Dcerpc: z.array(dcerpcEventSchema).optional(),
  Dhcp: z.array(dhcpEventSchema).optional(),
  Dns: z.array(dnsEventSchema).optional(),
  Fileinfo: z.array(fileinfoEventSchema).optional(),
  Flow: z.array(flowEventSchema).optional(),
  Ftp_Data: z.array(ftpDataEventSchema).optional(),
  Ftp: z.array(ftpEventSchema).optional(),
  Http: z.array(httpEventSchema).optional(),
  Krb5: z.array(krb5EventSchema).optional(),
  Mqtt: z.array(mqttEventSchema).optional(),
  Netflow: z.array(netflowEventSchema).optional(),
  Nfs: z.array(nfsEventSchema).optional(),
  Rdp: z.array(rdpEventSchema).optional(),
  Rfb: z.array(rfbEventSchema).optional(),
  Sip: z.array(sipEventSchema).optional(),
  Smb: z.array(smbEventSchema).optional(),
  Smb_Insights: z.array(smbInsightsEventSchema).optional(),
  Smtp: z.array(smtpEventSchema).optional(),
  Snmp: z.array(snmpEventSchema).optional(),
  Ssh: z.array(sshEventSchema).optional(),
  Stamus: z.array(stamusEventSchema).optional(),
  Tftp: z.array(tftpEventSchema).optional(),
  Tls: z.array(tlsEventSchema).optional(),
});

export type FlowEvents = z.infer<typeof flowEventsSchema>;